requirements.txt
# Core dependencies
openai==1.3.0
numpy==1.24.3
scikit-learn==1.3.0
python-dateutil==2.8.2
psutil==5.9.5
colorama==0.4.6

# Optional but recommended
python-dotenv==1.0.0

# Alternative if you prefer to use older OpenAI API syntax
# (uncomment line below and comment out openai==1.3.0 above)
# openai==0.28.1